Change the link to open a new issue using the template
I think GitHub has changed how to open a new issue using a template. When this page was written, a template could be selected from the new issue page. Now they have an interstitial, which the prior link was bypassing. Here, instead, we link directly to a new issue, pre-populated with the template.