additional cleanup & logging (closing - do not merge)
this branch was created in service of creating a one-time fix
for some broken records in our dynamodb. The scripts arent' really
helpful in a general-purpose sense and would otherwise require
cleanup and refactoring.  So I am committing the changes so we
can retrieve it from history but not merging into staging -JdV