use batchDelete for removeLock (see notes)
In the course of updating acquireLock#removeLock to use dynamodbClientService#batchDelete(),
I also went ahead and added typing for the for the batchDelete method. There was another
type warning in dynamodbClientService that was not introduced by me, but needed to be
addressed in order to commit, so i resorted to changing an untyped variable to type any[].
Apologies.